+++ CHANGES.rst
@@ -5,8 +5,7 @@ Changelog of threedi-modelchecker
 0.16 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added support for saving 2D initial waterlevel and groundwaterlevel in API. Note: uses first initial waterlevel (aggregation) resource found for 2D or groundwater.
 
 0.15 (2021-10-25)
 -----------------